Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int a,b,c;
-    scanf("%d %d %d",&a,&b,&c);
+    long int a,b,c;
+    scanf("%ld %ld %ld",&a,&b,&c);
     float e = (a+b+c)/3;
     printf("Average: %.2f",e);
     return 0;
